[wasm-backend] Initial WasmCodegen infrastructure
A WasmCodegen object will live in the Precompiler and contain the WasmModuleBuilder. It will be responsible for the whole Dart to Wasm compilation process (e.g. creating Wasm structs, translating the FlowGraphs).

This CL also implements the initial gathering of the classes and functions that will need to be compiled (a.k.a. hoisting) later on. This is necessary for allowing mutually recursive definitions.

Some refactoring has been employed as well when it comes to the interaction between Wasm code generation and the precompiler.


Tested manually
(the previous OutputWasm demo compiler pass has now moved to WasmCodegen::Demo()).

Change-Id: I7f10349cdb3e2518aec70ca50fbbbd3b0b127115
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/157320
Reviewed-by: Aske Simon Christensen <askesc@google.com>
Reviewed-by: Tess Strickland <sstrickl@google.com>
Commit-Queue: Andrei Constantinescu <andreicostin@google.com>