[wasm-backend] Dart classes and dispatch table calls for dart2wasm
In this CL we use the infrastructure we built using the Wasm GC
extension to compile code involving Dart classes. Points to note:

- We do not treat the case of IL constant instructions of object value;
- Instance calls work if they are static or if compiler optimizations
(in particular TFA) have managed to deduce the receiver type;
- For instance calls that need dynamic dispatch, we implemented support
for generating and populating a Wasm global dispatch table using the
Wasm table and element sections. By indexing into it, DispatchTableCalls
are implemented using the Wasm call_indirect instruction. Simple cases
of such dynamic dispatch are working, whereas more complex usage still
fail for an as of yet unknown reason.


Tested: Using run.js and the two test files: class-simple.dart and
class-dispatch.dart. In the first case, the code runs and prints the
right results. The second test involves a DispatchTableCall.

Change-Id: I894e2c98f0e9715fa093aa83eb9067837ba01160
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/165060
Reviewed-by: Aske Simon Christensen <askesc@google.com>